requirements.txt
fastapi
uvicorn
python-multipart
jinja2
opencv-python-headless
google-generativeai
numpy
easyocr